Allow event propagation on keyup
This was causing issues where keyup events were not getting bubbled to
the parent app. They were being stopped by the brush. This caused bugs
because we were using document events to show/hide the brush based on
the shift key. However we could not detect the shift key if the brush
was open. This was an edge case but an issue nonetheless.