Fix issue when selection is undefined
This was an issue with timeline slider. When the selection was
effectively nothing, then d3 was trying to access on null. This didn't
work of course. However, this issue emerged with the slider when both
handles were dragged all the way to one side, then tried to drag back.
Timeline slider does not need the custom logic of the graph selection
lib so its possible that the graph selection lib is causing issues.